Revert dependency between columns and serialization
